E5T2/feature/add ajax call
Add ajax call that removes the need to execute matrix_manip.php automatically

+++ index.php
@@ -1,6 +1,9 @@
 <?php 
 echo " ";   
-unlink("solution.json");
+$filename = "solution.json";
+if(file_exists($filename)){
+  unlink("solution.json");
+}
 ?>
 <!DOCTYPE html>
 <html lang="en">
